% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.poEMirtFit}
\alias{predict.poEMirtFit}
\title{Predict function}
\usage{
\method{predict}{poEMirtFit}(
  object,
  type = c("prob", "response"),
  return = c("array", "dataframe"),
  seed = NULL,
  ...
)
}
\arguments{
\item{object}{An object class of \code{poEMirtFit}}

\item{type}{A character, one of "prob" or "response". "prob" returns predicted probabilities of choices 
and "response" returns predicted value of Y.}

\item{return}{Type of returning object. "array" or "dataframe".}

\item{seed}{An integer of random seed. This works when \code{type = "response"}.}

\item{...}{Other arguments to \code{predict()}.}
}
\value{
3D array or dataframe
}
\description{
Predict function for `poEMirtFit`
}
