% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_poEMirt.R
\name{read_poEMirt}
\alias{read_poEMirt}
\title{Preparing for \code{poEMirt()}}
\usage{
read_poEMirt(
  data,
  responses,
  i,
  j,
  t = NULL,
  no_smooth = FALSE,
  matched_j = NULL,
  quiet = FALSE
)
}
\arguments{
\item{data}{A data.frame or tbl_df object.}

\item{responses}{Names of columns of responses. Last response category will be the baseline.}

\item{i}{A name of individual IDs column.}

\item{j}{A name of item IDs column.}

\item{t}{A name of time period IDs column (for `dynamic` model).}

\item{no_smooth}{A boot. If TRUE, the model does not estimate latent traits of i in missing years (for `dynamic` model). Default is FALSE.}

\item{matched_j}{A name of column that contains item IDs of matched item.}

\item{quiet}{A bool. Suppress messages (TRUE) or not. Default is FALSE.}
}
\value{
A \code{poEMirtData} object.
}
\description{
This function creates an input data for \code{poEMirt} models.
}
\examples{
\dontrun{
data("sim_data_dynamic")

# Convert into poEMirt-readable data
data <- read_poEMirt(
  data = sim_data_dynamic,
  responses = paste0('y', 1:5),
  i = "i",
  j = "j",
  t = "t"
)
}
}
