% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_data_dynamic_repeated}
\alias{sim_data_dynamic_repeated}
\title{Simulated data (dynamic and repeated items)}
\format{
A tbl_df with 200 individuals, 400 items (10 repeated items), 40 time-periods (10 items per time-period), and max(K_j) = 5. 
 \describe{
  \item{i}{Individual ID}
  \item{j}{Item ID}
  \item{t}{Time-period}
  \item{y1}{Response to k = 1}
  \item{y2}{Response to k = 2}
  \item{y3}{Response to k = 3}
  \item{y4}{Response to k = 4}
  \item{y5}{Response to k = 5}
}
}
\usage{
sim_data_dynamic_repeated
}
\description{
Simulated data (dynamic and repeated items)
}
\keyword{datasets}
