% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_data_static}
\alias{sim_data_static}
\title{Simulated data (static)}
\format{
a tbl_df with 500 individuals, 100 items, and max(K_j) = 5. 
 \describe{
  \item{i}{Individual ID}
  \item{j}{Item ID}
  \item{y1}{Response to k = 1}
  \item{y2}{Response to k = 2}
  \item{y3}{Response to k = 3}
  \item{y4}{Response to k = 4}
  \item{y5}{Response to k = 5}
}
}
\usage{
sim_data_static
}
\description{
Simulated data (static)
}
\keyword{datasets}
