% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.poEMirtBoot}
\alias{summary.poEMirtBoot}
\title{Summary function for \code{poEMirtBoot}}
\usage{
\method{summary}{poEMirtBoot}(object, parameter = c("alpha", "beta", "theta"), ci = 0.95, ...)
}
\arguments{
\item{object}{An object class of \code{poEMirtBoot}.}

\item{parameter}{A character or character vector, indicating what parameters ("alpha", "beta", "theta") you want to get.}

\item{ci}{A float (< 1). Confidence interval.}

\item{...}{Other arguments to \code{summary()}.}
}
\value{
A tibble dataframe.
}
\description{
\code{summary.poEMirtBoot} returns a dataframe of estimated parameters with confidence interval.
}
