% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.poEMirtData}
\alias{summary.poEMirtData}
\title{Summary function for \code{poEMirtData}}
\usage{
\method{summary}{poEMirtData}(object, ...)
}
\arguments{
\item{object}{An object class of \code{poEMirtFit}.}

\item{...}{Parameters to \code{summary()}.}
}
\description{
\code{summary.poEMirtFit} returns a summary of \code{poEMirtData} object.
}
