% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.poEMirtFit}
\alias{summary.poEMirtFit}
\title{Summary function for \code{poEMirtFit}}
\usage{
\method{summary}{poEMirtFit}(object, parameter = c("alpha", "beta", "theta"), ...)
}
\arguments{
\item{object}{An object class of \code{poEMirtFit}.}

\item{parameter}{A character or character vector, indicating what parameters ("alpha", "beta", "theta") you want to get.}

\item{...}{Other arguments to \code{summary()}.}
}
\value{
A tibble dataframe.
}
\description{
\code{summary.poEMirtFit} returns a dataframe of estimated parameters.
}
