% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.poEMirtGibbs}
\alias{summary.poEMirtGibbs}
\title{Summary function for \code{poEMirtGibbs}}
\usage{
\method{summary}{poEMirtGibbs}(object, parameter = c("alpha", "beta", "theta"), ci = 0.95, ...)
}
\arguments{
\item{object}{An object class of \code{poEMirtGibbs}.}

\item{parameter}{A character or character vector, indicating what parameters ("alpha", "beta", "theta") you want to get.}

\item{ci}{A float (< 1). Confidence interval.}

\item{...}{Parameters to \code{summary()}.}
}
\value{
A tibble dataframe.
}
\description{
\code{summary.poEMirtGibbs} returns a dataframe of estimated parameters with confidence interval.
}
