% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poEMirt.R
\name{poEMirt}
\alias{poEMirt}
\title{A fast EM Item Response Theory model for public opinion analysis}
\usage{
poEMirt(
  data,
  model = c("static", "dynamic"),
  constraint = NULL,
  alpha_fix = FALSE,
  theta_std = FALSE,
  init = NULL,
  priors = NULL,
  control = NULL
)
}
\arguments{
\item{data}{An object of \code{poEMirtData} via \code{read_poEMirt()}.}

\item{model}{A string, one of "static" or "dynamic".}

\item{constraint}{An integer scalar or vector (for dynamic model, the same length as the number of times), index of an individual i (the location of i) whose latent trait is always set positive.}

\item{alpha_fix}{A bool, whether fixes alpha of same repeated items or not. Default is FALSE.}

\item{theta_std}{A bool, whether standardizes theta or not. Default is FALSE}

\item{init}{A list, containing initial values (optional).
\itemize{
  \item \code{alpha} J x max(K)-1 matrix of alpha. Missing values must be 0.
  \item \code{beta} J x max(K)-1 matrix of beta. Missing values must be 0.
  \item \code{theta} I-length vector (for static model) or I x T matrix of theta (for dynamic model). Missing values must be 0.
}}

\item{priors}{a list, containing prior distributions (optional).
\itemize{
  \item \code{a0} A double or J x max(K)-1 matrix, prior mean of alpha. Default is 0.
  \item \code{A0} A double or J x max(K)-1 matrix, prior variance of alpha. Default is 1.
  \item \code{b0} A double or J x max(K)-1 matrix, prior mean of beta Default is 0.
  \item \code{B0} A double or J x max(K)-1 matrix, prior variance of beta. Default is 1.
  \item \code{m0} A double or I length vector, prior mean of theta_i0 for dynamic model. Default is 0. 
  \item \code{C0} A double or I length vector, prior variance of theta_i0 for dynamic model. Default is 1.
  \item \code{Delta} A double or I length vector, prior evolution variance of theta_it for dynamic model. Default is 0.01.
}}

\item{control}{A list of model controls.
\itemize{
  \item \code{compute_ll} A bool, whether compute log-likelihood for each iteration or not. Default is FALSE.
  \item \code{maxit} An integer, maximum number of iterations for EM. Default is 500.
  \item \code{tol} A double (< 1), convergence threshold. Default is 1e-6.
  \item \code{verbose} An integer, the function prints the status every `verbose`.
}}
}
\value{
A \code{poEMirtFit} object containing:
\describe{
  \item{parameter}{A list of parameters}
  \item{iteration}{The number of iterations (integer)}
  \item{converge}{Whether the model converged or not (boolean)}
  \item{convstat}{A matrix of convergence statistics}
  \item{info}{A list of model information}
}
}
\description{
Fit \code{poEMirt} models.
}
\examples{
\dontrun{
data("sim_data_dynamic")

# Convert into poEMirt-readable data
data <- read_poEMirt(
  data = sim_data_dynamic,
  responses = paste0('y', 1:5),
  i = "i",
  j = "j",
  t = "t"
)

# Fit the model
fit <- poEMirt(
  data = data,
  model = "dynamic",
  control = list(
    verbose = 10,
    constrant = 1
  )
)

# Summarize the result 
summary(fit, parameter = "theta")
}
}
