% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.poEMirtFit}
\alias{predict.poEMirtFit}
\title{Predict function for `poEMirtFit`}
\usage{
\method{predict}{poEMirtFit}(object, type = c("prob", "response"), ...)
}
\arguments{
\item{object}{An object class of \code{poEMirtFit}}

\item{type}{A character, one of "prob" or "response". "prob" returns predicted probabilities of choices 
and "response" returns predicted value of Y.}

\item{...}{Auxiliary parameters to \code{predict()}.}
}
\description{
Predict function for `poEMirtFit`
}
