% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_poEMirt.R
\name{read_poEMirt}
\alias{read_poEMirt}
\title{Preparing for \code{poEMirt()}}
\usage{
read_poEMirt(dataframe, responses, i, j, t = NULL, smooth = NULL)
}
\arguments{
\item{dataframe}{A data.frame or tbl_df object.}

\item{responses}{Names of columns of responses. Last response category for each j will be the baseline.}

\item{i}{A name of individual IDs column. Must be an integer column.}

\item{j}{A name of item IDs column. Must be an integer column.}

\item{t}{A name of time period IDs column (for `dynamic` model). Must be an integer column.}

\item{smooth}{A character, the smoothing method. 
\itemize{
  \item \code{"no"} No smoothing.
  \item \code{"default"} Using from the start time of i to T.
  \item \code{"finite"} Using from the start time of it to the end time.
 }}
}
\value{
A \code{poEMirtData} object.
}
\description{
This function creates an input data for \code{poEMirt} models.
}
\examples{
\dontrun{
data("sim_data_dynamic")

# Convert into poEMirt-readable data
data <- read_poEMirt(
  dataframe = sim_data_dynamic,
  responses = paste0('y', 1:5),
  i = "i",
  j = "j",
  t = "t"
)
}
}
